게시글 상세 조회 응답 수정
