+++ src/post/dtos/post.response.ts
@@ -41,6 +41,20 @@ class PostClothingDto {
   url: string;
 }
 
+class UserDto {
+  @ApiProperty({
+    example: 'nickname',
+    description: '사용자의 닉네임',
+  })
+  nickname: string;
+
+  @ApiProperty({
+    example: 'http://example.com/image.jpg',
+    description: '사용자의 프로필 사진 URL',
+  })
+  profilePictureUrl: string;
+}
+
 export class PostResponse {
   @ApiProperty({
     example: '게시물 번호',
@@ -95,6 +109,35 @@ export class PostResponse {
 }
 
 export class PatchPostResponse extends PostResponse {
-  @ApiProperty({ example: '2024-10-11T09:00:00.000Z', description: '수정 시각' })
+  @ApiProperty({
+    example: '2024-10-11T09:00:00.000Z',
+    description: '수정 시각',
+  })
   updatedAt: string;
+}
+
+export class PostDetailResponse extends PostResponse {
+  @ApiProperty({
+    type: UserDto,
+    description: '게시물 작성자 정보',
+  })
+  user: UserDto;
+
+  @ApiProperty({
+    example: 10,
+    description: '게시글에 달린 댓글 수입니다.',
+  })
+  commentCount: number;
+
+  @ApiProperty({
+    example: 5,
+    description: '게시글의 좋아요 수입니다.',
+  })
+  likeCount: number;
+
+  @ApiProperty({
+    example: false,
+    description: '현재 사용자가 게시물에 좋아요를 눌렀는지 여부',
+  })
+  isPostLike: boolean;
 }
