+++ src/post/post.controller.ts
@@ -29,7 +29,7 @@ import { PostRequest } from './dtos/post.request';
 import { BaseResponse } from 'src/common/response/dto';
 import { AuthGuard } from 'src/auth/guards/jwt.auth.guard';
 import { Request } from 'express';
-import { GetPostResponse } from './dtos/get-post.dto';
+import { PostDetailResponse } from './dtos/post.response';
 import { PatchPostRequest } from './dtos/post.request';
 import { PageOptionsDto } from './dtos/page-options.dto';
 import { PageDto } from './dtos/page.dto';
@@ -179,39 +179,36 @@ export class PostController {
   async getPost(
     @Param('postId') postId: number,
     @Req() req: Request,
-  ): Promise<BaseResponse<GetPostResponse>> {
-    const currentUserId = req.user.id;
-
+  ): Promise<BaseResponse<PostDetailResponse>> {
     const post = await this.postService.getPost(postId);
 
-    const postResponse: GetPostResponse = {
-      post: {
-        content: post.content,
-        createdAt: post.createdAt,
-        postImages: post.postImages
-          .filter((image) => image.status === 'activated')
-          .map((image) => ({
-            url: image.url,
-            orderNum: image.orderNum,
-          })),
-        postClothings: post.postClothings
-          .filter((postClothing) => postClothing.status === 'activated')
-          .map((postClothing) => ({
-            imageUrl: postClothing.clothing.imageUrl,
-            brandName: postClothing.clothing.brandName,
-            modelName: postClothing.clothing.modelName,
-            modelNumber: postClothing.clothing.modelNumber,
-            url: postClothing.clothing.url,
-          })),
-        likeCount: post.postLikes.length,
-        commentCount: post.postComments.length,
-        isPostLike: this.postService.checkIsPostLiked(post, currentUserId),
-        user: {
-          userId: post.user.id,
-          nickname: post.user.nickname,
-          profilePictureUrl: post.user.profilePictureUrl,
-        },
-        isPostWriter: post.user.id === currentUserId,
+    const postResponse: PostDetailResponse = {
+      postId: post.id,
+      userId: post.user.id,
+      createdAt: dayjs(post.createdAt).format('YYYY-MM-DDTHH:mm:ssZ'),
+      content: post.content,
+      postImages: post.postImages
+        .filter((image) => image.status === 'activated')
+        .map((image) => ({
+          url: image.url,
+          orderNum: image.orderNum,
+        })),
+      postClothings: post.postClothings
+        .filter((postClothing) => postClothing.status === 'activated')
+        .map((postClothing) => ({
+          imageUrl: postClothing.clothing.imageUrl,
+          brandName: postClothing.clothing.brandName,
+          modelName: postClothing.clothing.modelName,
+          modelNumber: postClothing.clothing.modelNumber,
+          url: postClothing.clothing.url,
+        })),
+      likeCount: post.postLikes.length,
+      commentCount: post.postComments.length,
+      isPostLike: this.postService.checkIsPostLiked(post, req.user.id),
+      isRepresentative: post.isRepresentative,
+      user: {
+        nickname: post.user.nickname,
+        profilePictureUrl: post.user.profilePictureUrl,
       },
     };
